rocker: Change COMP_STATUS to COMP_ERR to match changes in spec/device
Also pad desc so that comp_err is last two bytes of desc and completion gen bit
is last bit of those two bytes.  The resv padding desc is before comp_err.  The
reason for putting comp_err at very end of desc is we want comp_err to be the
last two bytes written in descriptor so if software is polling comp gen bit,
when it see gen bit change, it's safe to read other fields of desc.  Software
should read comp_err gen bit first, and then other desc fields.

Alos note that comp_err is 2 bytes, not 4.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>